Pass match accuracy to UI via MatchedBird DTO

+++ app/src/main/java/de/cdvost/jibjib/domain/interactors/web/dto/MatchResult.java
@@ -6,15 +6,15 @@ import de.cdvost.jibjib.repository.room.model.entity.Bird;
 
 public class MatchResult {
 
-    private double accuracy;
+    private float accuracy;
     private int id;
 
-    public MatchResult(double accuracy, int id) {
+    public MatchResult(float accuracy, int id) {
         this.accuracy = accuracy;
         this.id = id;
     }
 
-    public double getAccuracy() {
+    public float getAccuracy() {
         return accuracy;
     }
